+++ src/App/Types.hs
@@ -4,6 +4,7 @@ module App.Types (
   OverrideProject (..),
   ProjectMetadata (..),
   ProjectRevision (..),
+  MonorepoAnalysisOpts (..),
 ) where
 
 import Data.Text (Text)
@@ -28,6 +29,10 @@ data ProjectMetadata = ProjectMetadata
   , projectReleaseGroupRelease :: Maybe Text
   }
   deriving (Eq, Ord, Show)
+
+newtype MonorepoAnalysisOpts = MonorepoAnalysisOpts
+  { monorepoAnalysisType :: Maybe Text
+  }
 
 data ProjectRevision = ProjectRevision
   { projectName :: Text