+++ Changelog.md
@@ -1,5 +1,9 @@
 # Spectrometer Changelog
 
+## Unreleased 
+
+- Support HTTP endpoints for archive uploads ([#276](https://github.com/fossas/spectrometer/pull/276))
+
 ## v2.10.2
 
 - Fixes an issue where some `fossa` commands (including `fossa test`) would exit non-zero on success ([#278](https://github.com/fossas/spectrometer/pull/278)).