+++ Changelog.md
@@ -5,6 +5,12 @@
 - Support ReleaseGroup configuration ([#283](https://github.com/fossas/spectrometer/pull/283))
 - Support HTTP endpoints for archive uploads ([#276](https://github.com/fossas/spectrometer/pull/276))
 
+## v2.10.3
+
+- Deprecate `fossa vps` subcommands.
+- Add `--experimental-enable-monorepo` and other associated flags to `fossa analyze`, which enables experimental monorepo support.
+  Monorepo flags are only used with support from FOSSA engineering, and are experimental while we build robust support for monorepo projects with FOSSA partners.
+
 ## v2.10.2
 
 - Fixes an issue where some `fossa` commands (including `fossa test`) would exit non-zero on success ([#278](https://github.com/fossas/spectrometer/pull/278)).